requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pypdf
unstructured
python-dotenv
faiss-cpu
tiktoken